global batch

+++ train_elastic.py
@@ -94,7 +94,7 @@ def train():
         first_epoch = checkpoint["epoch"]
 
     # 创建数据加载器
-    batch_size = global_batch_size / world_size
+    batch_size = int(global_batch_size / world_size)
     dataset = KafkaDataset()
     # sampler = torch.utils.data.distributed.DistributedSampler(dataset, num_replicas=world_size,
     #                                                           rank=rank)
